Update typescript.md
The npm command mentioned in the guide doesn't work because there is no npm run strapi command that is defined in the project package.json scripts. Unlike yarn where you can run the command whether it is predefined or not. Running npx strapi ts:generate-types works. I also added a note for some issues with prettier that our team has ran into which may not apply to everyone.